whole code refactoring *Not Working* (compiling though)
no longer using global variables
GraphicsEngine is now the only class up to drawing
Interface was moved in Player to allow to keep it while switching player
include cleanning
some constness fix
reorganized Battle Map
using only enums class
using size_t for identifiers

removed draw() functions from every classes as only the graphicsEngine() is allowed to draw
removed events management

+++ src/common/structures/Vector.hh
@@ -40,7 +40,7 @@ public:
 
 
 # include <SFML/Graphics.hpp>
-typedef sf::Vector2<unsigned int> Coords; // FIXME (own [use ifndef SFML])
+typedef sf::Vector2<size_t> Coords; // FIXME (own [use ifndef SFML])
 typedef sf::Color Color; // FIXME (own [use ifndef SFML])
 // NOTE: this software uses SFML and does not try to hide it
 //   these defines are here to avoid
